feat: extract error enrichment to separate service
