+++ packages/ai-terminal-assistant/src/browser/ai-terminal-assistant-service.ts
@@ -14,24 +14,18 @@
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-only WITH Classpath-exception-2.0
 // *****************************************************************************
 import { inject, injectable, postConstruct } from '@theia/core/shared/inversify';
-import { AiTerminalSummaryAgent, ErrorDetail, ErrorLines, Summary } from './terminal-output-analysis-agent';
+import { AiTerminalSummaryAgent, ErrorDetail, Summary } from './terminal-output-analysis-agent';
 import { TerminalService } from '@theia/terminal/lib/browser/base/terminal-service';
 import { TerminalWidget } from '@theia/terminal/lib/browser/base/terminal-widget';
 import { TerminalWidgetImpl } from '@theia/terminal/lib/browser/terminal-widget-impl';
-import { Emitter, Event, URI } from '@theia/core';
+import { Emitter, Event } from '@theia/core';
 import { DebugSessionManager } from '@theia/debug/lib/browser/debug-session-manager';
-import { MonacoEditorService } from '@theia/monaco/lib/browser/monaco-editor-service';
-import * as monaco from '@theia/monaco-editor-core/esm/vs/editor/editor.api';
-import { LocalFileLinkProvider } from '@theia/terminal/lib/browser/terminal-file-link-provider';
 import { Widget } from '@theia/core/lib/browser';
 import { TaskTerminalWidgetManager } from '@theia/task/lib/browser/task-terminal-widget-manager';
 import { TaskWatcher } from '@theia/task/lib/common';
 import { TaskService } from '@theia/task/lib/browser/task-service';
-import { FileService } from '@theia/filesystem/lib/browser/file-service';
-import { FileSearchService } from '@theia/file-search/lib/common/file-search-service';
-import { WorkspaceService } from '@theia/workspace/lib/browser';
 import { AiTerminalAssistantPreferences } from './ai-terminal-assistant-preferences';
-import { IEditorDecorationsCollection } from '@theia/monaco-editor-core/esm/vs/editor/common/editorCommon';
+import { ErrorSourceResolutionService } from './error-source-resolution-service';
 
 export interface SummaryRequest {
     cwd: string;
@@ -75,17 +69,8 @@ export class SummaryServiceImpl implements SummaryService {
     @inject(DebugSessionManager)
     protected readonly debugSessionManager: DebugSessionManager;
 
-    @inject(MonacoEditorService)
-    protected readonly monacoEditorService: MonacoEditorService;
-
-    @inject(FileService)
-    protected readonly fileService: FileService;
-
-    @inject(LocalFileLinkProvider)
-    protected readonly fileLinkProvider: LocalFileLinkProvider;
-
-    @inject(FileSearchService)
-    protected readonly fileSearchService: FileSearchService;
+    @inject(ErrorSourceResolutionService)
+    protected readonly errorSourceResolutionService: ErrorSourceResolutionService;
 
     @inject(TaskTerminalWidgetManager)
     protected readonly taskTerminalWidgetManager: TaskTerminalWidgetManager;
@@ -95,9 +80,6 @@ export class SummaryServiceImpl implements SummaryService {
 
     @inject(TaskWatcher)
     protected readonly taskWatcher: TaskWatcher;
-
-    @inject(WorkspaceService)
-    protected readonly workspaceService: WorkspaceService;
 
     @inject(AiTerminalAssistantPreferences)
     protected readonly aiTerminalAssistantPreferences: AiTerminalAssistantPreferences;
@@ -121,8 +103,6 @@ export class SummaryServiceImpl implements SummaryService {
     readonly onCurrentTerminalBufferChanged: Event<void> = this.onCurrentTerminalBufferChangedEmitter.event;
 
     protected readonly activeTerminals = new Set<TerminalWidget>();
-    protected editorDecorations: monaco.editor.IEditorDecorationsCollection | undefined;
-    protected editorDisposables: monaco.IDisposable[] = [];
 
     protected _currentTerminal: TerminalWidget;
     get currentTerminal(): TerminalWidget {
@@ -150,8 +130,6 @@ export class SummaryServiceImpl implements SummaryService {
     }
 
     protected _isStandAlone: boolean;
-
-    protected decorationsCollection: IEditorDecorationsCollection;
 
     @postConstruct()
     protected initialize(): void {
@@ -323,7 +301,7 @@ export class SummaryServiceImpl implements SummaryService {
     async requestSummary(): Promise<void> {
         this.onSummaryRequestStartedEmitter.fire();
         const summary = await this.sendSummaryRequestForLastUsedTerminal();
-        this._currentSummary = summary ? await this.enrichErrorsWithFileContent(summary) : undefined;
+        this._currentSummary = summary ? await this.errorSourceResolutionService.enrichErrorsWithFileContent(summary) : undefined;
         this.onSummaryRequestFinishedEmitter.fire(this.currentSummary);
     }
 
@@ -355,97 +333,7 @@ export class SummaryServiceImpl implements SummaryService {
     }
 
     async openErrorInEditor(error: ErrorDetail): Promise<void> {
-        if (!error.file) {
-            throw new Error('Error does not contain file information.');
-        };
-        const terminal = this.terminalService.lastUsedTerminal;
-        if (!terminal) {
-            throw new Error('No active terminal found.');
-        }
-        const terminalLinks = await this.fileLinkProvider.provideLinks(`${error.file}${error.line ? ':' + error.line.toString() : ''}`, terminal);
-        const termminalLink = terminalLinks[0];
-        if (!termminalLink) {
-            throw new Error(`Could not find file link for ${error.file}`);
-        }
-        await termminalLink.handle();
-
-        const monacoEditor = this.monacoEditorService.getActiveCodeEditor();
-        if (monacoEditor) {
-            if (this.decorationsCollection) {
-                this.decorationsCollection.clear();
-            }
-            this.decorationsCollection = monacoEditor.createDecorationsCollection([{
-                range: new monaco.Range(error.line || 1, 1, error.line || 1, 1),
-                options: {
-                    className: 'error-line-decoration',
-                    description: 'error-line-decoration',
-                    hoverMessage: { value: `**Error**: ${error.fixSteps.join(' ')}` },
-                    isWholeLine: true,
-                }
-            }]);
-        }
-    };
-
-    protected async enrichErrorsWithFileContent(summary: Summary): Promise<Summary> {
-        const enrichedErrors = await Promise.all(summary.errors.map(async error => {
-            const errorLines = await this.getErrorLines(error);
-            return {
-                ...error,
-                errorLines,
-            };
-        }));
-        console.log('Enriched errors with file content:', enrichedErrors);
-        return {
-            ...summary,
-            errors: enrichedErrors,
-        };
-    }
-
-    protected async getErrorLines(error: ErrorDetail): Promise<ErrorLines | undefined> {
-        if (!error.file || !error.line) {
-            console.log('Error does not contain file or line information:', error);
-            return undefined;
-        }
-        const fileUri = await this.getFileUriFromError(error);
-        if (!fileUri) {
-            console.log('Could not find file URI for error:', error);
-            return undefined;
-        }
-        const uri: URI = new URI(fileUri);
-        const fileContent = await this.fileService.read(uri);
-        const lines = fileContent.value.split('\n');
-        console.log('Fetched file content for error lines:', lines);
-
-        const start = Math.max(0, error.line - 2);
-        const end = Math.min(lines.length, error.line + 1);
-        if (error.line && error.line > 0 && error.line <= lines.length) {
-            const errorLines = lines.slice(start, end);
-            // prefix each line with line number
-            const numberedLines = errorLines.map((line, index) => `${start + index + 1}: ${line}`);
-            return {
-                errorLines: numberedLines,
-                errorLinesStart: start + 1
-            };
-        }
-        return undefined;
-    }
-
-    protected async getFileUriFromError(error: ErrorDetail): Promise<string | undefined> {
-        if (!error.file) {
-            return undefined;
-        }
-        const searchTerm = error.file.replace(/`/g, '');
-        console.log('Searching for file URI with term:', searchTerm);
-        const roots = this.workspaceService.tryGetRoots().map(root => root.resource.toString());
-        const opts: FileSearchService.Options = {
-            rootUris: roots,
-            fuzzyMatch: true,
-            limit: 1,
-        };
-
-        const results = await this.fileSearchService.find(searchTerm, opts);
-
-        return results.length > 0 ? results[0] : undefined;
+        return this.errorSourceResolutionService.openErrorInEditor(error);
     }
 
     protected getRecentTerminalCommands(terminal: TerminalWidget): string[] {
